perubahan pu

+++ src/main/java/org/sim/umira/dtos/CostControl/ResponseProyekDto.java
@@ -5,14 +5,14 @@ import java.math.BigInteger;
 
 public class ResponseProyekDto<T> {
     public BigDecimal total_bk;
-    public Integer total_pu;
+    public BigInteger total_pu;
     public BigInteger current_mos;
     public BigInteger kerja_tambah;
     public BigInteger kerja_kurang;
     public BigInteger total_scurve;
     public T proyek;
     
-    public ResponseProyekDto(BigDecimal total_bk, Integer total_pu, BigInteger current_mos, BigInteger kerja_tambah, BigInteger kerja_kurang, BigInteger total_scurve, T proyek) {
+    public ResponseProyekDto(BigDecimal total_bk, BigInteger total_pu, BigInteger current_mos, BigInteger kerja_tambah, BigInteger kerja_kurang, BigInteger total_scurve, T proyek) {
         this.total_bk = total_bk;
         this.total_pu = total_pu;
         this.proyek = proyek;
